+++ backend/src/main/java/com/example/backend/localBoard/comment/controller/CommentQueryController.java
@@ -25,5 +25,5 @@ public class CommentQueryController {
         List<CommentResponseDTO> comments = commentQueryService.getBoardComment(localBoardId);
 
         return ResponseEntity.ok(comments);
-    };
+    }
 }
