+++ backend/src/main/java/com/example/backend/reward/rewardItems/controller/RewardItemsController.java
@@ -1,0 +1,54 @@
+package com.example.backend.reward.rewardItems.controller;
+
+import com.example.backend.reward.rewardItems.dto.request.RewardItemsRequestUpdateDTO;
+import com.example.backend.reward.rewardItems.dto.request.RewardRequestDTO;
+import com.example.backend.reward.rewardItems.dto.response.FindRewardDetailDTO;
+import com.example.backend.reward.rewardItems.dto.response.FindRewardListDTO;
+import com.example.backend.entity.reward.RewardItems;
+import com.example.backend.reward.rewardItems.service.RewardItemsService;
+import jakarta.validation.Valid;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.*;
+
+import java.util.List;
+
+@RestController
+@RequestMapping("/api/reward")
+public class RewardItemsController {
+    private final RewardItemsService rewardItemsService;
+
+    public RewardItemsController(RewardItemsService rewardItemsService) {
+        this.rewardItemsService = rewardItemsService;
+    }
+
+    @PostMapping("/register")
+    public ResponseEntity<String> rewardItemInsert(@Valid @RequestBody RewardRequestDTO dto){
+        rewardItemsService.rewardItemInsert(dto);
+        return new ResponseEntity<>("리워드 상품 등록 완료", HttpStatus.OK);
+    }
+
+    @GetMapping("/list/{rewardType}")
+    public ResponseEntity<List<FindRewardListDTO>> findRewardItemList(@PathVariable RewardItems.RewardType rewardType){
+        List<FindRewardListDTO> list = rewardItemsService.findRewardItemList(rewardType);
+        return new ResponseEntity<>(list,HttpStatus.OK);
+    }
+
+    @GetMapping("/detail/{id}")
+    public ResponseEntity<FindRewardDetailDTO> findRewardDetail(@PathVariable long id){
+        FindRewardDetailDTO dto  =  rewardItemsService.findRewardDetail(id);
+        return new ResponseEntity<>(dto,HttpStatus.OK);
+    }
+
+    @PutMapping("/update")
+    public ResponseEntity<String> rewardItemUpdate(@Valid @RequestBody RewardItemsRequestUpdateDTO dto){
+        rewardItemsService.rewardItemUpdate(dto);
+        return new ResponseEntity<>("게시물이 수정되었습니다.",HttpStatus.OK);
+    }
+
+    @DeleteMapping("/delete/{id}")
+    public ResponseEntity<String> rewardItemsDelete(@PathVariable long id){
+        rewardItemsService.rewardDeleteById(id);
+        return new ResponseEntity<>("게시물이 삭제되었습니다.",HttpStatus.OK);
+    }
+}
