+++ backend/src/main/java/com/example/backend/reward/rewardItems/repository/RewardRepository.java
@@ -1,0 +1,10 @@
+package com.example.backend.reward.rewardItems.repository;
+
+import com.example.backend.entity.reward.RewardItems;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface RewardRepository extends JpaRepository<RewardItems,Long>, RewardRepositoryCustom {
+
+}
