+++ backend/src/main/java/com/example/backend/reward/rewardItems/repository/RewardRepositoryCustom.java
@@ -1,0 +1,16 @@
+package com.example.backend.reward.rewardItems.repository;
+
+
+
+import com.example.backend.entity.reward.RewardItems;
+import org.springframework.data.repository.query.Param;
+import com.example.backend.reward.rewardItems.dto.response.FindRewardDetailDTO;
+import com.example.backend.reward.rewardItems.dto.response.FindRewardListDTO;
+
+import java.util.List;
+
+public interface RewardRepositoryCustom {
+    List<FindRewardListDTO> findRewardList(@Param("rewardType")RewardItems.RewardType rewardType);
+    FindRewardDetailDTO findRewardDetailById(@Param("id") long id);
+
+}
