+++ backend/src/main/java/com/example/backend/reward/rewardItmeImages/dto/request/RewardImagesUpdateDTO.java
@@ -1,0 +1,20 @@
+package com.example.backend.reward.rewardItmeImages.dto.request;
+
+import com.example.backend.entity.reward.RewardItemImages;
+import lombok.Getter;
+import lombok.Setter;
+
+@Setter
+@Getter
+public class RewardImagesUpdateDTO {
+    private long rewardItemId;
+    private String RewardItemsImgUrl;
+
+    public RewardItemImages toEntity(){
+        return RewardItemImages.builder()
+                .rewardItemId(rewardItemId)
+                .rewardItemsImgUrl(RewardItemsImgUrl)
+                .build();
+    }
+
+}
