+++ backend/src/main/java/com/example/backend/reward/rewardItmeImages/repository/RewardImageRepository.java
@@ -1,0 +1,10 @@
+package com.example.backend.reward.rewardItmeImages.repository;
+
+import com.example.backend.entity.reward.RewardItemImages;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface RewardImageRepository extends JpaRepository<RewardItemImages,Long>, RewardImageRepositoryCustom {
+    void deleteById(long id);
+}
