+++ backend/src/main/java/com/example/backend/reward/rewardItmeImages/repository/RewardImageRepositoryCustom.java
@@ -1,0 +1,5 @@
+package com.example.backend.reward.rewardItmeImages.repository;
+
+public interface RewardImageRepositoryCustom {
+    void deleteByAll(Long id);
+}
