+++ backend/src/main/java/com/example/backend/reward/userReward/repository/UserRewardRepositoryCustom.java
@@ -1,0 +1,8 @@
+package com.example.backend.reward.userReward.repository;
+
+import com.example.backend.reward.userReward.dto.response.UserRewardRecodeDTO;
+import org.springframework.data.repository.query.Param;
+
+public interface UserRewardRepositoryCustom {
+    UserRewardRecodeDTO findUserRewardById(@Param("id") long id);
+}
