--- backend/src/main/java/reward/rewardItems/repository/RewardRepository.java
@@ -1,10 +1,0 @@
-package reward.rewardItems.repository;
-
-import com.example.backend.entity.reward.RewardItems;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface RewardRepository extends JpaRepository<RewardItems,Long>, RewardRepositoryCustom {
-
-}
