--- backend/src/main/java/reward/rewardItmeImages/dto/request/RewardImagesUpdateDTO.java
@@ -1,20 +1,0 @@
-package reward.rewardItmeImages.dto.request;
-
-import com.example.backend.entity.reward.RewardItemImages;
-import lombok.Getter;
-import lombok.Setter;
-
-@Setter
-@Getter
-public class RewardImagesUpdateDTO {
-    private long rewardItemId;
-    private String RewardItemsImgUrl;
-
-    public RewardItemImages toEntity(){
-        return RewardItemImages.builder()
-                .rewardItemId(rewardItemId)
-                .rewardItemsImgUrl(RewardItemsImgUrl)
-                .build();
-    }
-
-}
