--- backend/src/main/java/reward/rewardItmeImages/repository/RewardImageRepositoryCustom.java
@@ -1,5 +1,0 @@
-package reward.rewardItmeImages.repository;
-
-public interface RewardImageRepositoryCustom {
-    void deleteByAll(Long id);
-}
