--- backend/src/main/java/reward/userReward/repository/UserRewardRepositoryCustom.java
@@ -1,8 +1,0 @@
-package reward.userReward.repository;
-
-import reward.userReward.dto.response.UserRewardRecodeDTO;
-import org.springframework.data.repository.query.Param;
-
-public interface UserRewardRepositoryCustom {
-    UserRewardRecodeDTO findUserRewardById(@Param("id") long id);
-}
